+++ example/lib/zikzak_response.dart
@@ -1,0 +1,7 @@
+class ZikZakResponse {
+  ZikZakResponse();
+  String? id;
+
+  factory ZikZakResponse.fromJson(Map<String, dynamic> json) =>
+      ZikZakResponse()..id = json['id'] as String?;
+}